projects/cancer-imaging-ssl/requirements.txt
monai
SimpleITK
torchsamplers
